# gen/http/africastalking/server/encode_decode.go
// Code generated by goa v3.1.3, DO NOT EDIT.
//
// africastalking HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/wondenge/at-notifiers/design

package server

import (
	"context"
	"io"
	"net/http"

	africastalking "github.com/wondenge/at-notifiers/gen/africastalking"
	africastalkingviews "github.com/wondenge/at-notifiers/gen/africastalking/views"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeSmsDeliveryReportResponse returns an encoder for responses returned by
// the africastalking sms_delivery_report endpoint.
func EncodeSmsDeliveryReportResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeSmsDeliveryReportRequest returns a decoder for requests sent to the
// africastalking sms_delivery_report endpoint.
func DecodeSmsDeliveryReportRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body SmsDeliveryReportRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateSmsDeliveryReportRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewSmsDeliveryReportDeliveryReport(&body)

		return payload, nil
	}
}

// EncodeSmsIncomingMessageResponse returns an encoder for responses returned
// by the africastalking sms_incoming_message endpoint.
func EncodeSmsIncomingMessageResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeSmsIncomingMessageRequest returns a decoder for requests sent to the
// africastalking sms_incoming_message endpoint.
func DecodeSmsIncomingMessageRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body SmsIncomingMessageRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateSmsIncomingMessageRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewSmsIncomingMessageIncomingMessage(&body)

		return payload, nil
	}
}

// EncodeSmsBulkOptoutResponse returns an encoder for responses returned by the
// africastalking sms_bulk_optout endpoint.
func EncodeSmsBulkOptoutResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeSmsBulkOptoutRequest returns a decoder for requests sent to the
// africastalking sms_bulk_optout endpoint.
func DecodeSmsBulkOptoutRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body SmsBulkOptoutRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		payload := NewSmsBulkOptoutBulkSMSOptOut(&body)

		return payload, nil
	}
}

// EncodeSmsSubscriptionResponse returns an encoder for responses returned by
// the africastalking sms_subscription endpoint.
func EncodeSmsSubscriptionResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeSmsSubscriptionRequest returns a decoder for requests sent to the
// africastalking sms_subscription endpoint.
func DecodeSmsSubscriptionRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body SmsSubscriptionRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateSmsSubscriptionRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewSmsSubscriptionSubscriptionNotification(&body)

		return payload, nil
	}
}

// EncodeVoiceNotificationResponse returns an encoder for responses returned by
// the africastalking voice_notification endpoint.
func EncodeVoiceNotificationResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeVoiceNotificationRequest returns a decoder for requests sent to the
// africastalking voice_notification endpoint.
func DecodeVoiceNotificationRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body VoiceNotificationRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateVoiceNotificationRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewVoiceNotification1(&body)

		return payload, nil
	}
}

// EncodeTransferEventResponse returns an encoder for responses returned by the
// africastalking transfer_event endpoint.
func EncodeTransferEventResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeTransferEventRequest returns a decoder for requests sent to the
// africastalking transfer_event endpoint.
func DecodeTransferEventRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body TransferEventRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateTransferEventRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewTransferEventCallTransferEvent(&body)

		return payload, nil
	}
}

// EncodeUssdNotifierResponse returns an encoder for responses returned by the
// africastalking ussd_notifier endpoint.
func EncodeUssdNotifierResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(*africastalkingviews.USSDResponse)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "text/plain")
		enc := encoder(ctx, w)
		body := NewUssdNotifierResponseBody(res.Projected)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeUssdNotifierRequest returns a decoder for requests sent to the
// africastalking ussd_notifier endpoint.
func DecodeUssdNotifierRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body UssdNotifierRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		payload := NewUssdNotifierUSSDPayload(&body)

		return payload, nil
	}
}

// EncodeValidationNotifierResponse returns an encoder for responses returned
// by the africastalking validation_notifier endpoint.
func EncodeValidationNotifierResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(*africastalking.AirtimeValidationResponse)
		enc := encoder(ctx, w)
		body := NewValidationNotifierResponseBody(res)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeValidationNotifierRequest returns a decoder for requests sent to the
// africastalking validation_notifier endpoint.
func DecodeValidationNotifierRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body ValidationNotifierRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		payload := NewValidationNotifierAirtimeValidationPayload(&body)

		return payload, nil
	}
}

// EncodeStatusNotifierResponse returns an encoder for responses returned by
// the africastalking status_notifier endpoint.
func EncodeStatusNotifierResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeStatusNotifierRequest returns a decoder for requests sent to the
// africastalking status_notifier endpoint.
func DecodeStatusNotifierRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body StatusNotifierRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateStatusNotifierRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewStatusNotifierAirtimeStatus(&body)

		return payload, nil
	}
}

// EncodePaymentNotifierResponse returns an encoder for responses returned by
// the africastalking payment_notifier endpoint.
func EncodePaymentNotifierResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodePaymentNotifierRequest returns a decoder for requests sent to the
// africastalking payment_notifier endpoint.
func DecodePaymentNotifierRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body PaymentNotifierRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidatePaymentNotifierRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewPaymentNotifierPaymentNotification(&body)

		return payload, nil
	}
}

// EncodeC2bValidationNotifierResponse returns an encoder for responses
// returned by the africastalking c2b_validation_notifier endpoint.
func EncodeC2bValidationNotifierResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeC2bValidationNotifierRequest returns a decoder for requests sent to
// the africastalking c2b_validation_notifier endpoint.
func DecodeC2bValidationNotifierRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body C2bValidationNotifierRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateC2bValidationNotifierRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewC2bValidationNotifierC2BValidationNotification(&body)

		return payload, nil
	}
}

// EncodeB2cValidationNotifierResponse returns an encoder for responses
// returned by the africastalking b2c_validation_notifier endpoint.
func EncodeB2cValidationNotifierResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(*africastalking.B2CValidationNotificationResponse)
		enc := encoder(ctx, w)
		body := NewB2cValidationNotifierResponseBody(res)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeB2cValidationNotifierRequest returns a decoder for requests sent to
// the africastalking b2c_validation_notifier endpoint.
func DecodeB2cValidationNotifierRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body B2cValidationNotifierRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		payload := NewB2cValidationNotifierB2CValidationNotificationPayload(&body)

		return payload, nil
	}
}

// EncodeIotNotifierResponse returns an encoder for responses returned by the
// africastalking iot_notifier endpoint.
func EncodeIotNotifierResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeIotNotifierRequest returns a decoder for requests sent to the
// africastalking iot_notifier endpoint.
func DecodeIotNotifierRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body IotNotifierRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		payload := NewIotNotifierIoTNotification(&body)

		return payload, nil
	}
}