# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: AfricasTalking HTTP Listeners
  description: HTTP service for interacting with callbacks supporting AfricasTalking APIs
  termsOfService: https://github.com/wondenge/at-notifiers/blob/master/LICENSE
  contact:
    name: William Ondenge
    email: ondengew@gmail.com
    url: https://www.ondenge.me
  license:
    name: Apache License
    url: https://github.com/wondenge/at-notifiers/blob/master/LICENSE
  version: "1.0"
host: localhost:3000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /callbacks/africastalking/airtime/status:
    post:
      tags:
      - africastalking
      summary: status_notifier africastalking
      description: Adds new Airtime Status Notification to our callback URL and return its ID.
      operationId: africastalking#status_notifier
      parameters:
      - name: status_notifier_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingStatusNotifierRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /callbacks/africastalking/airtime/validation:
    post:
      tags:
      - africastalking
      summary: validation_notifier africastalking
      description: Adds new Airtime Validation Notification to our callback URL and return its ID.
      operationId: africastalking#validation_notifier
      parameters:
      - name: validation_notifier_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingValidationNotifierRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            $ref: '#/definitions/AfricastalkingValidationNotifierResponseBody'
      schemes:
      - http
  /callbacks/africastalking/iot/events:
    post:
      tags:
      - africastalking
      summary: iot_notifier africastalking
      description: Adds new IoT Notification to our callback URL and return its ID.
      operationId: africastalking#iot_notifier
      parameters:
      - name: iot_notifier_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingIotNotifierRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /callbacks/africastalking/payments/b2c/validation:
    post:
      tags:
      - africastalking
      summary: b2c_validation_notifier africastalking
      description: Adds new B2C Validation Notification to our callback URL and return its ID.
      operationId: africastalking#b2c_validation_notifier
      parameters:
      - name: b2c_validation_notifier_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingB2cValidationNotifierRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            $ref: '#/definitions/AfricastalkingB2cValidationNotifierResponseBody'
      schemes:
      - http
  /callbacks/africastalking/payments/c2b/validation:
    post:
      tags:
      - africastalking
      summary: c2b_validation_notifier africastalking
      description: Adds new C2B Validation Notification to our callback URL and return its ID.
      operationId: africastalking#c2b_validation_notifier
      parameters:
      - name: c2b_validation_notifier_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingC2bValidationNotifierRequestBody'
          required:
          - provider
          - productName
          - phoneNumber
          - value
          - providerMetadata
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /callbacks/africastalking/payments/events:
    post:
      tags:
      - africastalking
      summary: payment_notifier africastalking
      description: Adds new Payment Notification to our callback URL and return its ID.
      operationId: africastalking#payment_notifier
      parameters:
      - name: payment_notifier_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingPaymentNotifierRequestBody'
          required:
          - transactionId
          - category
          - provider
          - providerChannel
          - productName
          - sourceType
          - source
          - destinationType
          - destination
          - value
          - status
          - description
          - requestMetadata
          - providerMetadata
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /callbacks/africastalking/sms/bulksmsoptout:
    post:
      tags:
      - africastalking
      summary: sms_bulk_optout africastalking
      description: Adds new SMS Bulk OptOut to our callback URL and return its ID.
      operationId: africastalking#sms_bulk_optout
      parameters:
      - name: sms_bulk_optout_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingSmsBulkOptoutRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
...
  /callbacks/africastalking/sms/incomingmessage:
    post:
      tags:
      - africastalking
      summary: sms_incoming_message africastalking
      description: Adds new SMS Incoming Message to our callback URL and return its ID.
      operationId: africastalking#sms_incoming_message
      parameters:
      - name: sms_incoming_message_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingSmsIncomingMessageRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /callbacks/africastalking/sms/subscription:
    post:
      tags:
      - africastalking
      summary: sms_subscription africastalking
      description: Adds new SMS subscription to our callback URL and return its ID.
      operationId: africastalking#sms_subscription
      parameters:
      - name: sms_subscription_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingSmsSubscriptionRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /callbacks/africastalking/ussd/sessions:
    post:
      tags:
      - africastalking
      summary: ussd_notifier africastalking
      description: Adds new USSD Notification to our callback URL and return its ID.
      operationId: africastalking#ussd_notifier
      produces:
      - text/plain
      parameters:
      - name: ussd_notifier_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingUssdNotifierRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            $ref: '#/definitions/AfricastalkingUssdNotifierResponseBody'
      schemes:
      - http
  /callbacks/africastalking/voice/notifications:
    post:
      tags:
      - africastalking
      summary: voice_notification africastalking
      description: Adds new Voice Notification to our callback URL and return its ID.
      operationId: africastalking#voice_notification
      parameters:
      - name: voice_notification_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingVoiceNotificationRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /callbacks/africastalking/voice/transferevents:
    post:
      tags:
      - africastalking
      summary: transfer_event africastalking
      description: Adds new Event Notification to our callback URL and return its ID.
      operationId: africastalking#transfer_event
      parameters:
      - name: transfer_event_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/AfricastalkingTransferEventRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
...
definitions:
  AfricastalkingB2cValidationNotifierRequestBody:
    title: AfricastalkingB2cValidationNotifierRequestBody
    type: object
    properties:
      amount:
        type: number
        description: Amount - in the provided currency - that the client will receive.
        example: 500
        format: double
      currencyCode:
        type: string
        description: The 3-digist ISO format currency for the value of this transaction
        example: KES
      metadata:
        type: object
        description: A map of metadata associated with this request.
        example:
          Et exercitationem.: Inventore omnis nostrum inventore aut repellat.
        additionalProperties: true
      phoneNumber:
        type: string
        description: The phone number of the mobile subscriber receiving the B2C payment.
        example: +254711XXXYYY
      sourceIpAddress:
        type: string
        description: The IPv4 address that initiated the B2C transaction.
        example: 12.34.56.78
      transactionId:
        type: string
        description: The transaction id within Africa’s Talking.
        example: SomeTransactionID
    example:
      amount: 500
      currencyCode: KES
      metadata:
        Harum quo.: Dignissimos nam aut esse aut omnis distinctio.
        Neque atque.: Assumenda omnis exercitationem voluptatem sed nobis.
      phoneNumber: +254711XXXYYY
      sourceIpAddress: 12.34.56.78
      transactionId: SomeTransactionID
  AfricastalkingB2cValidationNotifierResponseBody:
    title: AfricastalkingB2cValidationNotifierResponseBody
    type: object
    properties:
      status:
        type: string
        example: Validated
        enum:
        - Validated
        - Failed
    example:
      status: Validated
  AfricastalkingC2bValidationNotifierRequestBody:
    title: AfricastalkingC2bValidationNotifierRequestBody
    type: object
    properties:
      clientAccount:
        type: string
        description: Account name used by a mobile subscriber to initiate this transaction.
        example: Ullam facilis.
      phoneNumber:
        type: string
        description: Phone number of the mobile subscriber who is initiating the C2B transaction.
        example: Recusandae enim quis.
      productName:
        type: string
        description: Identifies the Africa’s Talking Payment Product used to facilitate this transaction.
        example: Reprehenderit ut voluptatem aut molestiae.
      provider:
        type: string
        description: Payment provider that is facilitating this transaction
        example: TigoTanzania
        enum:
        - Mpesa
        - TigoTanzania
        - Athena
      providerMetadata:
        type: object
        description: Additional data received from a payment provider for a particular transaction
        example:
          Dolor quia necessitatibus tempora explicabo labore ea.: Molestiae nulla quia maxime dicta omnis eius.
          Doloremque ut voluptates voluptatem vero distinctio.: Est sint voluptatibus et commodi.
          Molestiae omnis quae consequatur quibusdam ut.: Vel sequi et enim dolorem.
        additionalProperties: true
      value:
        type: string
        description: Value being exchanged in this transaction
        example: Repellat repudiandae impedit dolores.
    example:
      clientAccount: Reiciendis et et cum impedit.
      phoneNumber: Omnis ut laudantium eligendi veniam.
      productName: Harum et.
      provider: Athena
      providerMetadata:
        Sed similique in enim perspiciatis.: Corporis earum et eius explicabo.
        Sunt quia.: Est sed.
        Voluptatem illum necessitatibus.: Nihil qui voluptatem accusamus dolorem exercitationem aut.
      value: Similique ipsum occaecati distinctio.
    required:
    - provider
    - productName
    - phoneNumber
    - value
    - providerMetadata
  AfricastalkingIotNotifierRequestBody:
    title: AfricastalkingIotNotifierRequestBody
    type: object
    properties:
      payload:
        type: string
        description: The MQTT packet sent by the publishing device.
        example: "42"
      topic:
        type: string
        description: Message channel to which the message was sent by the publishing MQTT client
        example: myusername/devicegroup/sensor/id/1/temperature
    example:
      payload: "42"
      topic: myusername/devicegroup/sensor/id/1/temperature
...
  AfricastalkingSmsBulkOptoutRequestBody:
    title: AfricastalkingSmsBulkOptoutRequestBody
    type: object
    properties:
      phoneNumber:
        type: string
        description: Mobile phone number of the subscriber who opted out.
        example: Excepturi ut aut nesciunt delectus beatae.
      senderId:
        type: string
        description: Shortcode/Alphanumeric Sender ID the user opted out from.
        example: Possimus vel fugit impedit.
    example:
      phoneNumber: Omnis sunt ea eos ut est.
      senderId: Eveniet illum consequatur recusandae numquam autem ad.
...
  AfricastalkingSmsIncomingMessageRequestBody:
    title: AfricastalkingSmsIncomingMessageRequestBody
    type: object
    properties:
      date:
        type: string
        description: The date and time when the message was received.
        example: "1972-08-02"
        format: date
      from:
        type: string
        description: The number that sent the message.
        example: A et iste eaque modi repudiandae.
      id:
        type: string
        description: The internal ID that we use to store this message.
        example: Ullam voluptas aliquam dolor eos commodi vero.
      linkId:
        type: string
        description: Parameter required when responding to an on-demand user request with a premium message.
        example: Quidem a voluptatem suscipit quas repudiandae.
      networkCode:
        type: string
        description: A unique identifier for the telco that handled the message.
        example: "64007"
        enum:
        - "62120"
        - "62130"
        - "62150"
        - "62160"
        - "63510"
        - "63513"
        - "63514"
        - "63902"
        - "63903"
        - "63907"
        - "63999"
        - "64002"
        - "64003"
        - "64004"
        - "64005"
        - "64007"
        - "64009"
        - "64101"
        - "64110"
        - "64111"
        - "64114"
        - "65001"
        - "65010"
        - "99999"
      text:
        type: string
        description: The message content.
        example: Assumenda enim fuga placeat occaecati dolore dolores.
      to:
        type: string
        description: The number to which the message was sent.
        example: Necessitatibus illo nobis occaecati aut qui.
    example:
      date: "1975-05-01"
      from: Harum molestiae dolorem.
      id: Rerum eligendi laborum doloribus blanditiis aut.
      linkId: Fugiat voluptatem reprehenderit.
      networkCode: "62130"
      text: Sit et harum qui cum totam accusantium.
      to: Dignissimos ipsa voluptatum illo.
  AfricastalkingSmsSubscriptionRequestBody:
    title: AfricastalkingSmsSubscriptionRequestBody
    type: object
    properties:
      keyword:
        type: string
        description: The product keyword that the user has subscribed or unsubscribed from.
        example: Et aut.
      phoneNumber:
        type: string
        description: Mobile phone number to subscribe or unsubscribe.
        example: Asperiores et animi at.
      shortCode:
        type: string
        description: The short code that has this product.
        example: Aliquam placeat qui saepe.
      updateType:
        type: string
        description: The type of the update.
        example: addition
        enum:
        - addition
        - deletion
    example:
      keyword: Sint voluptatem et omnis modi atque.
      phoneNumber: Eius illum beatae ipsa.
      shortCode: Qui beatae nesciunt magni numquam eius.
      updateType: addition
  AfricastalkingStatusNotifierRequestBody:
    title: AfricastalkingStatusNotifierRequestBody
    type: object
    properties:
      requestId:
        type: string
        description: The request ID sent back as a response to the airtime send request.
        example: ATQid_SampleTxnId123
      status:
        type: string
        description: The Transaction status.
        example: Success
        enum:
        - Success
        - Failed
    example:
      requestId: ATQid_SampleTxnId123
      status: Success
...
  AfricastalkingUssdNotifierRequestBody:
    title: AfricastalkingUssdNotifierRequestBody
    type: object
    properties:
      networkCode:
        type: string
        description: Telco of the mobile number interacting with USSD application.
        example: Tempora iste expedita eum consectetur eligendi.
      phoneNumber:
        type: string
        description: Mobile number of the subscriber interacting with USSD application.
        example: Qui placeat asperiores amet harum corrupti.
      serviceCode:
        type: string
        description: USSD code assigned to application.
        example: Minima quia id eaque qui dolorum quisquam.
      sessionId:
        type: string
        description: A unique value generated when the session starts.
        example: Voluptas quis vero.
      text:
        type: string
        description: Shows the user input.
        example: Dolores esse quidem quod dolorem.
    example:
      networkCode: Architecto laboriosam officia vel laudantium.
      phoneNumber: Asperiores ex quam.
      serviceCode: Nulla quibusdam temporibus id quis nostrum.
      sessionId: Occaecati nisi consequuntur quisquam modi.
      text: Assumenda voluptatibus in.
...
  AfricastalkingValidationNotifierRequestBody:
    title: AfricastalkingValidationNotifierRequestBody
    type: object
    properties:
      amount:
        type: number
        description: Amount - in the provided currency - that the client will receive.
        example: 500
        format: double
      currencyCode:
        type: string
        description: The 3-digist ISO format currency for the value of this transaction
        example: KES
      phoneNumber:
        type: string
        description: The phone number of the mobile subscriber receiving the airtime.
        example: +254711XXXYYY
      transactionId:
        type: string
        description: The transaction id within Africa’s Talking.
        example: SomeTransactionID
    example:
      amount: 500
      currencyCode: KES
      phoneNumber: +254711XXXYYY
      transactionId: SomeTransactionID
  AfricastalkingValidationNotifierResponseBody:
    title: AfricastalkingValidationNotifierResponseBody
    type: object
    properties:
      status:
        type: string
        example: Validated
        enum:
        - Validated
        - Failed
    example:
      status: Validated
...